requirements.txt
rctclient>=0.0.3
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
slowapi>=0.1.9
pyyaml>=6.0.1
httpx>=0.27.0
shutil; python_version < "3.0"  # stdlib, listed for clarity
